+++ R/app_config.R
@@ -33,6 +33,10 @@ init_config <- function(project) {
     config <<- yaml::read_yaml(config_file)
     verify_config_file(config)
 
+    if (is.null(config$light_version)) {
+        config$light_version <- FALSE
+    }
+
     verify_app_cache(project)
 
     help_config <<- yaml::read_yaml(help_file)