+++ R/mod_query.R
@@ -57,7 +57,7 @@ mod_query_ui <- function(id) {
                     closable = FALSE,
                     width = 12,
                     shinycssloaders::withSpinner(
-                        plotly::plotlyOutput(ns("plot_mc_stacked_type"))
+                        plotOutput(ns("plot_mc_stacked_type"))
                     )
                 )
             ),
@@ -349,7 +349,7 @@ mod_query_server <- function(id, dataset, metacell_types, cell_type_colors, gene
                 sliderInput(ns("min_edge_size"), label = "Min edge length", min = 0, max = 0.3, value = min_edge_length(dataset()), step = 0.001)
             })
 
-            output$plot_mc_stacked_type <- plot_type_predictions_bar(dataset)
+            output$plot_mc_stacked_type <- plot_type_predictions_bar(dataset, metacell_types, cell_type_colors)
 
             output$gene_metadata_cell_type_selector <- cell_type_selector(dataset, ns, id = "gene_metadata_cell_type", label = "Cell types", selected = "all", cell_type_colors = cell_type_colors)