+++ R/plot_type_prediction_bar.R
@@ -1,6 +1,5 @@
-
-plot_type_predictions_bar <- function(dataset) {
-    plotly::renderPlotly({
+plot_type_predictions_bar <- function(dataset, metacell_types, cell_type_colors) {
+    renderPlot({
         df_fracs <- get_mc_data(dataset(), "query_atlas_cell_type_fracs")
         req(!is.null(df_fracs))
         req(has_atlas(dataset()))
@@ -20,7 +19,7 @@ plot_type_predictions_bar <- function(dataset) {
             mutate(metacell = factor(metacell, levels = rownames(fracs_mat)[ord$rows])) %>%
             mutate(type = factor(type, levels = colnames(fracs_mat)[ord$cols]))
 
-        p <- df_fracs %>%
+        p_fracs <- df_fracs %>%
             ggplot(aes(x = metacell, y = fraction, fill = type, customdata = metacell)) +
             geom_col() +
             scale_fill_manual(name = "", values = atlas_colors) +
@@ -35,14 +34,19 @@ plot_type_predictions_bar <- function(dataset) {
             xlab("Metacell") +
             guides(fill = "none")
 
+        cell_type_colors <- cell_type_colors() %>%
+            select(cell_type, color) %>%
+            deframe()
 
-        fig <- plotly::ggplotly(
-            p,
-            source = "type_prediction_bar"
-        ) %>%
-            sanitize_plotly_buttons() %>%
-            plotly::hide_legend()
+        p_annot <- metacell_types() %>% 
+            mutate(metacell = factor(metacell, levels = rownames(fracs_mat)[ord$rows])) %>%
+            ggplot(aes(x = metacell, y = 1, fill = cell_type)) +
+                geom_raster() +
+                scale_fill_manual(name = "", values = cell_type_colors) +
+                theme_void() + 
+                ggplot2::theme(panel.border = ggplot2::element_rect(size = 0.2, color = "black", fill = NA))
 
-        return(fig)
-    })
+        p <- cowplot::insert_xaxis_grob(p_fracs, p_annot, grid::unit(0.05, "null"), position = "bottom")
+        cowplot::ggdraw(p)
+    }, res = 92)
 }