+++ R/plot_mc_mc_gene.R
@@ -154,7 +154,7 @@ render_mc_mc_gene_diff_table <- function(input, output, session, ns, dataset, mc
                     mc_mc_gene_scatter_df() %>%
                         filter(col != "gray") %>%
                         arrange(diff) %>%
-                        select(Gene = gene, `Diff (log2)` = diff, `P-value` = pval, any_of(c("Systematic", "Disjoined"))) %>%
+                        select(Gene = gene, `Diff (log2)` = diff, `P-value` = pval, any_of(c("Disjoined"))) %>%
                         mutate(GeneCards = glue("<a href='{link}' target='_blank'>Open</a>", link = paste0("https://www.genecards.org/cgi-bin/carddisp.pl?gene=", Gene))),
                     selection = "single",
                     escape = FALSE,