+++ NEWS.md
@@ -5,6 +5,7 @@
 * Added a 'rename' modal to the 'Annotate' tab.
 * Fixed order of negative correlations in the left panels. Fixes #127.
 * Added an option to use 'gene_names' parameter with atlas projection. 
+* Use the name of the dataset as title by default.
 
 # MCView 0.2.5