+++ NEWS.md
@@ -1,6 +1,7 @@
 # MCView 0.2.6
 
 * Use new `metacells` version. Note that old `anndata` objects will not work with this version without conversion.
+* Added an option to bundle a 'light' version without the option to change the genes on heatmaps.
 * Fixed resolution of heatmap plots.
 * Added a 'rename' modal to the 'Annotate' tab.
 * Fixed order of negative correlations in the left panels. Fixes #127.